+++ apps/web/lib/webhook/transform.ts
@@ -1,6 +1,9 @@
-import { toCamelCase } from "@dub/utils";
+import { nanoid, toCamelCase } from "@dub/utils";
 import type { Webhook } from "@prisma/client";
+import { WebhookTrigger } from "../types";
 import { LinkSchema } from "../zod/schemas/links";
+import { webhookPayloadSchema } from "../zod/schemas/webhooks";
+import { WEBHOOK_EVENT_ID_PREFIX } from "./constants";
 import { clickSchema, leadSchema, saleSchema } from "./schemas";
 
 interface TransformWebhookProps
@@ -80,5 +83,34 @@ export const transformSale = (data: any) => {
     link: {
       id: sale.linkId,
     },
+  });
+};
+
+// Transform the payload to the format expected by the webhook
+export const prepareWebhookPayload = (trigger: WebhookTrigger, data: any) => {
+  switch (trigger) {
+    case "link.created":
+    case "link.updated":
+    case "link.deleted":
+      data = transformLink(data);
+      break;
+    case "link.clicked":
+      data = transformClick(data);
+      break;
+    case "lead.created":
+      data = transformLead(data);
+      break;
+    case "sale.created":
+      data = transformSale(data);
+      break;
+    default:
+      break;
+  }
+
+  return webhookPayloadSchema.parse({
+    id: `${WEBHOOK_EVENT_ID_PREFIX}${nanoid(25)}`,
+    data: data,
+    event: trigger,
+    createdAt: new Date().toISOString(),
   });
 };
